refactor: correct file names and project structure
