--- rs-react-app/src/App.tsx
@@ -1,14 +1,0 @@
-import './App.css';
-import { Footer } from './components/footer/Footer';
-import { Header } from './components/header/Header';
-import { Main } from './components/main/Main';
-
-export const App = () => {
-  return (
-    <>
-      <Header />
-      <Main />
-      <Footer />
-    </>
-  );
-};
