+++ rs-react-app/src/App/App.tsx
@@ -1,0 +1,14 @@
+import './App.css';
+import { Footer } from '../components/Footer/Footer';
+import { Header } from '../components/Header/Header';
+import { Main } from '../components/Main/Main';
+
+export const App = () => {
+  return (
+    <>
+      <Header />
+      <Main />
+      <Footer />
+    </>
+  );
+};
