+++ rs-react-app/src/components/ErrorBoundary/ErrorBoundary.tsx
