+++ rs-react-app/src/main.tsx
@@ -1,8 +1,8 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
-import { App } from './App.tsx';
-import { ErrorBoundary } from './components/errorBoundeary/ErrorBoundary.tsx';
+import { App } from './App/App.tsx';
+import { ErrorBoundary } from './components/ErrorBoundary/ErrorBoundary.tsx';
 import { BrowserRouter } from 'react-router-dom';
 
 const rootElement = document.getElementById('root');
