+++ rs-react-app/src/types/characterTypes.ts
